feat: Allow to configure Actions Overview Elements Size - MEED-7358 - Meeds-io/MIPs#155 (#1710)
