+++ portlets/src/main/resources/locale/portlet/Challenges_en.properties
@@ -584,3 +584,18 @@ gamification.programs.overviewSettings.sortByCreatedDate=Creation date
 gamification.programs.overviewSettings.sortByModifiedDate=Modification date
 gamification.programs.overviewSettings.sortItemsBy=Sort items by:
 gamification.programs.overviewSettings.itemsCountLabel=Select number of items to list
+gamification.programs.overviewSettings.editTooltip=Edit settings
+
+gamification.rules.overviewSettings.title=Edit Actions List
+gamification.rules.overviewSettings.save=Save
+gamification.rules.overviewSettings.cancel=Cancel
+gamification.rules.overviewSettings.sortByScore=Score
+gamification.rules.overviewSettings.sortByTitle=Title
+gamification.rules.overviewSettings.sortByCreatedDate=Creation date
+gamification.rules.overviewSettings.sortByModifiedDate=Modification date
+gamification.rules.overviewSettings.sortItemsBy=Sort items by:
+gamification.rules.overviewSettings.itemsCountLabel=Select number of items to list
+gamification.rules.overviewSettings.actionsEndingSoon=Actions ending soon
+gamification.rules.overviewSettings.actionsToDoFirst=Actions to do first
+gamification.rules.overviewSettings.actionsAvailable=Actions Available
+gamification.rules.overviewSettings.actionsSoonAvailable=Actions Soon Available
