+++ portlets/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -429,7 +429,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>extensionRegistry</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
       <depends>
         <module>eCharts</module>
@@ -492,7 +492,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>commonVueComponents</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
     </module>
   </portlet>
@@ -526,7 +526,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>ruleComponents</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
     </module>
   </portlet>
@@ -560,7 +560,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>ruleComponents</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
     </module>
   </portlet>
@@ -597,7 +597,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>ruleComponents</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
     </module>
   </portlet>
@@ -826,7 +826,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
         <module>extensionRegistry</module>
       </depends>
       <depends>
-        <module>rulesOverviewWidget</module>
+        <module>usersLeaderboardCommon</module>
       </depends>
     </module>
   </portlet>
@@ -880,6 +880,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
       <module>vue</module>
     </depends>
     <depends>
+      <module>vuetify</module>
+    </depends>
+    <depends>
       <module>commonVueComponents</module>
     </depends>
     <depends>
@@ -887,6 +890,18 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     </depends>
     <depends>
       <module>ruleComponents</module>
+    </depends>
+    <depends>
+      <module>connectorExtensions</module>
+    </depends>
+    <depends>
+      <module>gamificationConnectorEvents</module>
+    </depends>
+    <depends>
+      <module>gamificationUserConnectors</module>
+    </depends>
+    <depends>
+      <module>translationField</module>
     </depends>
   </module>
 
@@ -1095,50 +1110,6 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     </depends>
     <depends>
       <module>ruleExtensions</module>
-    </depends>
-  </module>
-
-  <module>
-    <name>rulesOverviewWidget</name>
-    <script>
-      <minify>false</minify>
-      <path>/js/rulesOverviewWidget.bundle.js</path>
-    </script>
-    <depends>
-      <module>ruleComponents</module>
-    </depends>
-    <depends>
-      <module>connectorExtensions</module>
-    </depends>
-    <depends>
-      <module>gamificationConnectorEvents</module>
-    </depends>
-    <depends>
-      <module>gamificationUserConnectors</module>
-    </depends>
-    <depends>
-      <module>translationField</module>
-    </depends>
-    <depends>
-      <module>vue</module>
-    </depends>
-    <depends>
-      <module>vuetify</module>
-    </depends>
-    <depends>
-      <module>eXoVueI18n</module>
-    </depends>
-    <depends>
-      <module>commonVueComponents</module>
-    </depends>
-    <depends>
-      <module>extensionRegistry</module>
-    </depends>
-    <depends>
-      <module>gamificationCommon</module>
-    </depends>
-    <depends>
-      <module>usersLeaderboardCommon</module>
     </depends>
   </module>
