+++ portlets/src/main/webapp/vue-app/usersLeaderboardCommon/components/RulesOverviewListDrawer.vue
@@ -116,6 +116,7 @@ export default {
     activeRules: [],
     upcomingRules: [],
     endingRules: [],
+    spaceId: eXo.env.portal.spaceId,
     weekInMs: 604800000,
   }),
   computed: {
@@ -213,6 +214,7 @@ export default {
         programStatus: 'ENABLED',
         dateFilter: 'STARTED_WITH_END',
         programId: this.programId,
+        spaceId: !this.programId && this.spaceId?.length && [this.spaceId] || null,
         offset: 0,
         limit: 2 * this.page,
         sortBy: 'endDate',
@@ -228,6 +230,7 @@ export default {
         programStatus: 'ENABLED',
         dateFilter: 'STARTED',
         programId: this.programId,
+        spaceId: !this.programId && this.spaceId?.length && [this.spaceId] || null,
         offset: 0,
         limit: this.limit,
         sortBy: 'createdDate',
@@ -246,6 +249,7 @@ export default {
         programStatus: 'ENABLED',
         dateFilter: 'UPCOMING',
         programId: this.programId,
+        spaceId: !this.programId && this.spaceId?.length && [this.spaceId] || null,
         offset: 0,
         limit: 2 * this.page,
         sortBy: 'startDate',
