+++ portlets/src/main/webapp/vue-app/usersLeaderboardCommon/initComponents.js
@@ -26,10 +26,9 @@ import ProfileRealizationItem from './components/ProfileRealizationItem.vue';
 import ProfileAchievementsDrawer from './components/ProfileAchievementsDrawer.vue';
 import ProgramSelectorDrawer from './components/ProgramSelectorDrawer.vue';
 import RulesOverviewListDrawer from './components/RulesOverviewListDrawer.vue';
-import RulesOverviewReducedList from './components/RulesOverviewReducedList.vue';
-import RulesOverviewFullList from './components/RulesOverviewFullList.vue';
-import RulesOverviewSpaceList from './components/RulesOverviewSpaceList.vue';
 import RulesOverviewItem from './components/RulesOverviewItem.vue';
+import RulesOverviewWidget from './components/RulesOverviewWidget.vue';
+import RulesOverviewEmptyWidget from './components/RulesOverviewEmptyWidget.vue';
 
 const components = {
   'users-leaderboard-tabs': Tabs,
@@ -42,10 +41,9 @@ const components = {
   'users-leaderboard-profile-achievements-drawer': ProfileAchievementsDrawer,
   'users-leaderboard-program-selector-drawer': ProgramSelectorDrawer,
   'gamification-rules-overview-list-drawer': RulesOverviewListDrawer,
-  'gamification-rules-overview-reduced-list': RulesOverviewReducedList,
-  'gamification-rules-overview-full-list': RulesOverviewFullList,
-  'gamification-rules-overview-space-list': RulesOverviewSpaceList,
   'gamification-rules-overview-item': RulesOverviewItem,
+  'gamification-rules-overview-widget': RulesOverviewWidget,
+  'gamification-rules-overview-empty-widget': RulesOverviewEmptyWidget,
 };
 
 for (const key in components) {
